# app/static/swagger.yaml
swagger: "2.0"
info:
  title: "Mechanic Shop API"
  description: "API documentation for the mechanic shop management system"
  version: "1.0.0"
host: "127.0.0.1:5000"
schemes:
  - "http"
consumes:
  - "application/json"
produces:
  - "application/json"
securityDefinitions:
  bearerAuth:
    type: apiKey
    name: Authorization
    in: header
...
paths:
...
  /customers:
    get:
      tags: [Customers]
      summary: "Get paginated customers"
      description: "Returns a paginated list of customers."
      parameters:
        - in: query
          name: page
          type: integer
        - in: query
          name: per_page
          type: integer
      responses:
        200:
          description: "List of customers"
          schema:
            type: array
            items:
              $ref: "#/definitions/Customer"

    post:
      tags: [Customers]
      summary: "Create a customer"
      description: "Admins and mechanics only."
      security:
        - bearerAuth: []
      parameters:
        - in: body
          name: customer
          required: true
          schema:
            $ref: "#/definitions/CustomerCreatePayload"
      responses:
        201:
          description: "Customer created"
          schema:
            $ref: "#/definitions/CustomerCreateResponse"
        400:
          description: "Validation error"
        409:
          description: "Email exists"
  /customers/my-tickets:
    get:
      tags: [Customers]
      summary: "Get tickets for authenticated customer"
      description: "Returns all tickets belonging to the logged-in customer."
      security:
        - bearerAuth: []
      responses:
        200:
          description: "List of tickets"
          schema:
            type: array
            items:
              $ref: "#/definitions/CustomerTicket"
  /customers/{customer_id}:
    get:
      tags: [Customers]
      summary: "Get a customer"
      description: "Retrieve a customer by ID."
      parameters:
        - in: path
          name: customer_id
          required: true
          type: integer
      responses:
        200:
          description: "Customer found"
          schema:
            $ref: "#/definitions/Customer"
        404:
          description: "Customer not found"

    put:
      tags: [Customers]
      summary: "Update a customer"
      description: "Admins and mechanics only."
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: customer_id
          required: true
          type: integer
        - in: body
          name: customer
          required: true
          schema:
            $ref: "#/definitions/CustomerUpdatePayload"
      responses:
        200:
          description: "Customer updated"
          schema:
            $ref: "#/definitions/CustomerUpdateResponse"
        400:
          description: "Validation error"
        404:
          description: "Customer not found"
        409:
          description: "Email exists"

    delete:
      tags: [Customers]
      summary: "Delete a customer"
      description: "Admin only."
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: customer_id
          required: true
          type: integer
      responses:
        200:
          description: "Customer deleted"
          schema:
            $ref: "#/definitions/CustomerDeleteResponse"
        404:
          description: "Customer not found"

  # VEHICLES
...
  /tickets:
    get:
      tags: [Service Tickets]
      summary: "Get all service tickets"
      description: "Returns all service tickets."
      responses:
        200:
          description: "List of tickets"
          schema:
            type: array
            items:
              $ref: "#/definitions/ServiceTicket"

    post:
      tags: [Service Tickets]
      summary: "Create a service ticket"
      description: "Admins and mechanics only."
      security:
        - bearerAuth: []
      parameters:
        - in: body
          name: ticket
          required: true
          schema:
            $ref: "#/definitions/TicketCreatePayload"
      responses:
        201:
          description: "Ticket created"
          schema:
            $ref: "#/definitions/TicketCreateResponse"
        400:
          description: "Validation error"
  /tickets/{ticket_id}:
    get:
      tags: [Service Tickets]
      summary: "Get a service ticket"
      description: "Retrieve a ticket by ID."
      parameters:
        - in: path
          name: ticket_id
          required: true
          type: integer
      responses:
        200:
          description: "Ticket found"
          schema:
            $ref: "#/definitions/ServiceTicket"
        404:
          description: "Ticket not found"

    put:
      tags: [Service Tickets]
      summary: "Update a service ticket"
      description: "Admins and mechanics only."
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: ticket_id
          required: true
          type: integer
        - in: body
          name: ticket
          required: true
          schema:
            $ref: "#/definitions/TicketUpdatePayload"
      responses:
        200:
          description: "Ticket updated"
          schema:
            $ref: "#/definitions/TicketUpdateResponse"
        400:
          description: "Validation error"
        404:
          description: "Ticket not found"

    delete:
      tags: [Service Tickets]
      summary: "Delete a service ticket"
      description: "Admin only."
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: ticket_id
          required: true
          type: integer
      responses:
        200:
          description: "Ticket deleted"
          schema:
            $ref: "#/definitions/TicketDeleteResponse"
        404:
          description: "Ticket not found"
  /tickets/{ticket_id}/add_part:
    post:
      tags: [Service Tickets]
      summary: "Add a part to a ticket"
      description: "Admins and mechanics only."
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: ticket_id
          required: true
          type: integer
        - in: body
          name: part
          required: true
          schema:
            $ref: "#/definitions/AddPartPayload"
      responses:
        200:
          description: "Part added"
          schema:
            $ref: "#/definitions/AddPartResponse"
        404:
          description: "Ticket or part not found"
  /tickets/{ticket_id}/edit:
    put:
      tags: [Service Tickets]
      summary: "Add or remove mechanics from a ticket"
      description: "Admins and mechanics only."
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: ticket_id
          required: true
          type: integer
        - in: query
          name: add_ids
          type: array
          items:
            type: integer
        - in: query
          name: remove_ids
          type: array
          items:
            type: integer
      responses:
        200:
          description: "Mechanics updated"
          schema:
            $ref: "#/definitions/EditMechanicsResponse"
        404:
          description: "Ticket not found"

  # ASSIGNMENTS
...
definitions:
...
  Customer:
    type: object
    properties:
      id:
        type: integer
      name:
        type: string
      email:
        type: string
      phone:
        type: string
      created_at:
        type: string
  CustomerTicket:
    type: object
    properties:
      ticket_id:
        type: integer
      vehicle_id:
        type: integer
      status:
        type: string
      description:
        type: string
  CustomerCreatePayload:
    type: object
    required: [name, email]
    properties:
      name:
        type: string
      email:
        type: string
      phone:
        type: string
  CustomerCreateResponse:
    type: object
    properties:
      message:
        type: string
      customer:
        $ref: "#/definitions/Customer"
  CustomerUpdatePayload:
    type: object
    properties:
      name:
        type: string
      email:
        type: string
      phone:
        type: string
  CustomerUpdateResponse:
    type: object
    properties:
      message:
        type: string
      customer:
        $ref: "#/definitions/Customer"
  CustomerDeleteResponse:
    type: object
    properties:
      message:
        type: string

  # VEHICLES
...
  ServiceTicket:
    type: object
    properties:
      id:
        type: integer
      customer_id:
        type: integer
      vehicle_id:
        type: integer
      status:
        type: string
        enum: [open, in_progress, completed]
      description:
        type: string
      created_at:
        type: string
  TicketCreatePayload:
    type: object
    required: [customer_id, vehicle_id, description]
    properties:
      customer_id:
        type: integer
      vehicle_id:
        type: integer
      description:
        type: string
  TicketCreateResponse:
    type: object
    properties:
      message:
        type: string
      ticket:
        $ref: "#/definitions/ServiceTicket"
  TicketUpdatePayload:
    type: object
    properties:
      status:
        type: string
        enum: [open, in_progress, completed]
      description:
        type: string
  TicketUpdateResponse:
    type: object
    properties:
      message:
        type: string
      ticket:
        $ref: "#/definitions/ServiceTicket"
  TicketDeleteResponse:
    type: object
    properties:
      message:
        type: string
  AddPartPayload:
    type: object
    required: [part_id, quantity]
    properties:
      part_id:
        type: integer
      quantity:
        type: integer
  AddPartResponse:
    type: object
    properties:
      message:
        type: string
  EditMechanicsResponse:
    type: object
    properties:
      message:
        type: string

  # ASSIGNMENTS
...